feat: extract tuic-core, add tokio-quiche backend, runtime backend switch
Split the TUIC implementation into a shared, backend-agnostic core and two
interchangeable QUIC backends, then wire runtime backend selection into
tuic-server.

tuic-core (new):
- Backend-agnostic TUIC protocol primitives shared by both backends:
  wire codecs (header/command/address), the pure decode helpers used on the
  hot path, and the UDP fragment-reassembly state machine.
- ProtoError decoupled from quinn.

wind-tuic:
- Removed the placeholder quiche backend; re-exports tuic_core::proto so
  existing wind_tuic::proto::* paths keep working. Keeps the quinn-specific
  ClientProtoExt and the quinn-coupled UdpStream send path.

wind-tuiche (new): tokio-quiche backend
- Real TUIC server over tokio-quiche's ApplicationOverQuic worker
  (crates/wind-tuiche/src/driver.rs): auth, TCP CONNECT relay (bridged to
  wind-core via a channel-backed duplex stream), native UDP relay over
  RFC 9221 datagrams using the shared fragment buffer, heartbeat, dissociate.
- Pinned to tokio-quiche 0.18 to keep serde_with/quiche aligned with the
  workspace. Known limitation: quiche does not expose the RFC 5705 keying
  material exporter, so TUIC token verification is UUID-gated (documented).

tuic-server:
- New [backend] config section: backend.mode = "quinn" | "quiche" with
  per-backend tuning under backend.quinn / backend.quiche.
- create_inbound dispatches on backend.mode (ServerInbound::{Tuic, Tuiche}).
- Legacy top-level [quic] section and flat keys migrate into backend.quinn,
  so existing configs keep working.

tests:
- New crates/tuic-tests/tests/protocol_tests.rs (14 tests): wire-format
  stability, decode-helper/codec agreement, malformed-input handling,
  TargetAddr<->Address conversions, UDP reassembly via tuic-core, and the
  [backend] config surface.
- Backend-config unit tests in tuic-server.

+++ crates/tuic-core/src/proto/header.rs
@@ -134,7 +134,7 @@ mod test {
 		writer.send(header.clone()).await?;
 		let mut buffer = writer.into_inner();
 		let full_len = buffer.len();
-		let mut half_b = buffer.split_off(full_len / 2 as usize);
+		let mut half_b = buffer.split_off(full_len / 2_usize);
 		let mut half_a = buffer;
 		{
 			let mut reader = FramedRead::new(half_a.as_slice(), HeaderCodec);
